refactor(daemon): extract sandbox identity constants out of the k8s driver
Move the module-level pure constants, resolvers and the launch-timeout error
out of k8s/driver.ts into k8s/sandbox-identity.ts. driver.ts re-exports every
moved symbol, so no importer changes are needed.
