+++ packages/daemon/src/k8s/sandbox-identity.ts
@@ -1,0 +1,42 @@
+import type { ShimCapability } from '../shim/protocol.js'
+import type { Sandbox } from './sandbox-api.js'
+
+/** Label domain the claim controller must be configured to allow. */
+export const AC_LABEL_ORG = 'agentconnect.md/org'
+export const AC_LABEL_AGENT = 'agentconnect.md/agent'
+
+/** Where agent-sandbox records the pod a Sandbox is backed by — the only pod reference the daemon can read, since `SandboxStatus` carries none and the Pod API is outside its Role. */
+export const SANDBOX_POD_NAME_ANNOTATION = 'agents.x-k8s.io/pod-name'
+
+/** The pod backing this Sandbox: the adopted warm-pool pod, or the Sandbox's own name. */
+export function resolvePodName(sandbox: Sandbox): string | undefined {
+  const adopted = sandbox.metadata?.annotations?.[SANDBOX_POD_NAME_ANNOTATION]
+  return adopted && adopted.length > 0 ? adopted : sandbox.metadata?.name
+}
+
+/** The first routable address reported for the pod backing this Sandbox. */
+export function resolvePodIp(sandbox: Sandbox): string | undefined {
+  for (const entry of sandbox.status?.podIPs ?? []) {
+    const ip = typeof entry === 'string' ? entry : entry.ip
+    if (ip?.trim()) return ip.trim()
+  }
+  return undefined
+}
+
+/** Capabilities a runtime launch receives. Narrow by construction: a launch gets exactly
+ *  what the channels it uses require, so a future capability is an explicit decision. */
+export const RUNTIME_GRANTS: ShimCapability[] = ['acp', 'materialize', 'exec', 'read', 'tunnel']
+
+/** A probe sandbox runs no runtime and touches no workspace, so it gets the one channel it uses
+ *  and nothing else. Granting `probe` to every launch instead would hand each agent's runtime an
+ *  authority it never exercises — which the direct-connect grant test exists to catch. */
+export const PROBE_GRANTS: ShimCapability[] = ['probe']
+
+/** A launch stage that ran out of time. Typed, because a missed target and a broken cluster are
+ *  different operational stories and telling them apart by error text is a liability. */
+export class LaunchTimeoutError extends Error {
+  constructor(message: string) {
+    super(message)
+    this.name = 'LaunchTimeoutError'
+  }
+}
